requirements.txt
psycopg2-binary
django-environ
django-extensions
Pillow
django-autoslug
django-crispy-forms
django-debug-toolbar 